Trello #118 - Modificar graficos
