+++ src/components/dashboard/GraficoPrestadoresPorEspecialidad.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { PieChart, Pie, Cell, ResponsiveContainer, Tooltip, Legend } from 'recharts';
+import { BarChart, Bar, XAxis, YAxis, CartesianGrid, ResponsiveContainer, Tooltip, Cell } from 'recharts';
 import type { Prestador } from '../../types/prestadores';
 import './GraficoPrestadoresPorEspecialidad.css';
 
@@ -24,42 +24,18 @@ const GraficoPrestadoresPorEspecialidad: React.FC<Props> = ({ prestadores }) => 
     });
 
     return Array.from(conteo.entries())
-      .map(([name, value]) => ({ name, value }))
-      .sort((a, b) => b.value - a.value);
+      .map(([nombre, cantidad]) => ({ nombre, cantidad }))
+      .sort((a, b) => b.cantidad - a.cantidad);
   };
 
   const data = contarPorEspecialidad();
 
   // Paleta de colores
-  const COLORS = [
+  const colors = [
     '#4B81D8', '#8196c7', '#ff9d0a', '#5cb85c', '#f0ad4e', 
     '#d9534f', '#5bc0de', '#292b2c', '#0275d8', '#5a6268',
     '#17a2b8', '#6c757d', '#28a745', '#ffc107', '#dc3545'
   ];
-
-  // Custom label para mostrar porcentaje
-  const renderCustomLabel = ({ cx, cy, midAngle, innerRadius, outerRadius, percent }: any) => {
-    if (percent < 0.05) return null; // No mostrar label si es menos del 5%
-    
-    const RADIAN = Math.PI / 180;
-    const radius = innerRadius + (outerRadius - innerRadius) * 0.5;
-    const x = cx + radius * Math.cos(-midAngle * RADIAN);
-    const y = cy + radius * Math.sin(-midAngle * RADIAN);
-
-    return (
-      <text 
-        x={x} 
-        y={y} 
-        fill="white" 
-        textAnchor={x > cx ? 'start' : 'end'} 
-        dominantBaseline="central"
-        fontSize="14"
-        fontWeight="600"
-      >
-        {`${(percent * 100).toFixed(0)}%`}
-      </text>
-    );
-  };
 
   if (data.length === 0) {
     return (
@@ -83,21 +59,24 @@ const GraficoPrestadoresPorEspecialidad: React.FC<Props> = ({ prestadores }) => 
       </div>
       <div className="grafico-pie-content">
         <ResponsiveContainer width="100%" height={320}>
-          <PieChart>
-            <Pie
-              data={data}
-              cx="50%"
-              cy="50%"
-              labelLine={false}
-              label={renderCustomLabel}
-              outerRadius={110}
-              fill="#8884d8"
-              dataKey="value"
-            >
-              {data.map((_, index) => (
-                <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
-              ))}
-            </Pie>
+          <BarChart 
+            data={data}
+            margin={{ top: 5, right: 10, left: 5, bottom: 70 }}
+          >
+            <CartesianGrid strokeDasharray="3 3" stroke="#dee2e6" />
+            <XAxis 
+              dataKey="nombre" 
+              angle={-45}
+              textAnchor="end"
+              height={120}
+              interval={0}
+              tick={{ fill: '#606060', fontSize: 12 }}
+            />
+            <YAxis 
+              allowDecimals={false}
+              tick={{ fill: '#606060', fontSize: 12 }}
+              label={{ value: 'Cantidad', angle: -90, position: 'insideLeft', fill: '#424242' }}
+            />
             <Tooltip 
               contentStyle={{ 
                 backgroundColor: 'white', 
@@ -105,16 +84,19 @@ const GraficoPrestadoresPorEspecialidad: React.FC<Props> = ({ prestadores }) => 
                 borderRadius: '8px',
                 boxShadow: '0 2px 4px rgba(0,0,0,0.1)'
               }}
+              labelStyle={{ color: '#424242', fontWeight: 600 }}
               itemStyle={{ color: '#606060' }}
             />
-            <Legend 
-              verticalAlign="bottom" 
-              height={36}
-              wrapperStyle={{ paddingTop: '20px' }}
-              iconType="circle"
-              formatter={(value: string) => <span style={{ color: '#424242', fontSize: '0.9rem' }}>{value}</span>}
-            />
-          </PieChart>
+            <Bar 
+              dataKey="cantidad" 
+              radius={[8, 8, 0, 0]}
+              maxBarSize={80}
+            >
+              {data.map((_, index) => (
+                <Cell key={`cell-${index}`} fill={colors[index % colors.length]} />
+              ))}
+            </Bar>
+          </BarChart>
         </ResponsiveContainer>
       </div>
     </div>
